first version of socket_prod.py
This has working subroutines for tests used in production.  Also includes a working gui.  The connection from the GUI to running the tests as a result of the GUI was not yet completed or tested yet, due to COVID-19 shutdown.  This will be the basis for converting to larpix-v2 asic testing as well.